+++ epserde/tests/fail/drop_encase_deref.rs
@@ -5,16 +5,11 @@
  */
 
 use epserde::prelude::*;
-use std::io::Cursor;
 use std::ops::Deref;
 
 fn main() {
     let v = vec![0, 10, 20, 30, 40];
-
-    let mut buffer = Vec::new();
-    unsafe { v.serialize(&mut buffer).unwrap() };
-    let cursor = Cursor::new(&buffer);
-    let mem_case = unsafe { Vec::<i32>::read_mem(cursor, buffer.len()).unwrap() };
+    let mem_case = MemCase::encase(v);
     let d = mem_case.deref();
     drop(mem_case);
     let _d = d;